+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# v0.1.7
+-   #15 Mac OS X build of decode_barcodes
+
 # v0.1.6
 - #12 bin to scripts
 - #10 Rename images based on value of barcodes